Se trabajo con simulacion de carga

+++ MavenIW3/iw3/src/main/java/com/tpfinal/iw3/model/DetalleOrden.java
@@ -46,7 +46,7 @@ public class DetalleOrden {
     private float caudal;
 
     @Column(name = "estampa_tiempo", nullable = false)
-    private float estampaTiempo;
+    private Long estampaTiempo;  // Timestamp Unix en milisegundos
